argos/backend/requirements.txt
django~=4.2.3
docxtpl~=0.16.7
pytest~=7.4.0
openpyxl~=3.1.2
djangorestframework~=3.14.0
psycopg2-binary~=2.9.6
psycopg~=3.1.9
python-docx
django-factory~=0.11
factory-boy~=3.2.1
Faker~=19.1.0
django-cors-headers~=4.2.0